requirements.txt
appnope==0.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
decorator==4.1.2
Django==1.11.7
django-cors-headers==2.1.0
django-extensions==1.9.7
django-redis==4.8.0
djangorestframework==3.7.3
future==0.16.0
idna==2.6
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
pyexcel-io==0.2.1
pyexcel-xls==0.2.1
Pygments==2.2.0
python-slugify==1.2.4
pytz==2017.3
redis==2.10.6
requests==2.18.4
rets==0.4.2
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.2
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7
xlrd==1.0.0
xlwt-future==0.8.0
xmltodict==0.11.0
